Move browser code to separate module
